Normalize the Shopping, Trading, and Finding modules
The modules now allow the appid, certid, and devid to be set
when they are initialized, as well as via a yaml config file.

I also removed the username/password authorization option for
the trading API, since it seems it has been depricated in favor
of the user token.

Since the trading API doesn't allow http access, only https, if
a user tries to pass in https=False to the trading init function,
they will now get an error, and it will be set to true. The same
goes for the other modules and their respective https defaults.

Normalize the Shopping, Trading, and Finding modules

The modules now allow the appid, certid, and devid to be set
when they are initialized, as well as via a yaml config file.

I also removed the username/password authorization option for
the trading API, since it seems it has been depricated in favor
of the user token.

Since the trading API doesn't allow http access, only https, if
a user tries to pass in https=False to the trading init function,
they will now get an error, and it will be set to true. The same
goes for the other modules and their respective https defaults.

Code/spacing cleanup

Adding the certid and appid to the Shopping module

Removing HTTP requirement for Finding api since HTTPS is supported