Add files argument to execute
In the Trading api, there is currently no way to call UploadSiteHostedPictures using local files. This requires building a multipart request where the first part of the request is the XML and the second part is the binary MIME attachment.

Adding a files argument to the execute command and passing that through to the Request constructor allows the body to be constructed as multipart/form-data. We delete the 'Content-Type' header since the Trading api's build_request_header method sets it to 'text/xml' and the Request constructor (via requests > models.py > PreparedRequest.prepare_body) will not replace the 'Content-Type' header if it exists.

The value of the 'files' argument is as described in the requests module. See 'POST a Multipart-Encoded File' in http://docs.python-requests.org/en/latest/user/quickstart/ 

API behaviour is undefined if more that one file is passed in.